Change calibration resolution to offer decimal step Add the date unit at the end of the curve upper the axis Change mStepMinWidth parameter to accord it with the other axis, it was doing trouble. Change connection in ModelView for mMinEdit and mMaxEdit, i do it because appeare date format-unit Modify connection in MainWindow with mcmcFinished Add updateFormatDate in MainWindow

+++ src/model/EventKnown.cpp
@@ -92,7 +92,7 @@ void EventKnown::updateValues(double tmin, double tmax, double step)
     {
         case eFixed:
         {
-            for(int t=tmin; t<=tmax; t+=step)
+            for(double t=tmin; t<=tmax; t+=step)
                 mValues[t] = 0.f;
             mValues[mFixed] = 1.f;
             break;
@@ -101,7 +101,7 @@ void EventKnown::updateValues(double tmin, double tmax, double step)
         {
             if(mUniformStart < mUniformEnd)
             {
-                for(int t=tmin; t<=tmax; t+=step)
+                for(double t=tmin; t<=tmax; t+=step)
                 {
                     double v = (t > mUniformStart && t <= mUniformEnd) ? 1 / (mUniformEnd - mUniformStart) : 0;
                     mValues[t] = v;
@@ -114,7 +114,7 @@ void EventKnown::updateValues(double tmin, double tmax, double step)
     }
     if(mValues.size() == 0)
     {
-        for(int t=tmin; t<=tmax; t+=step)
+        for(double t=tmin; t<=tmax; t+=step)
             mValues[t] = 0.f;
     }
 }
